Had to change almost everything. The program need to separate bewtween different organizations and into that separation, split the datasets between tweets and replies. I added to the header a combobox to select the organization and now the index page works perfectly fine. Its time to finish the Thesaurus with this conditionals, then sentiment and finally, add the clouds of words to the page

+++ src/main/java/delco/twitter/scraping/controller/TweetController.java
@@ -90,8 +90,13 @@ public class TweetController {
     }
 
     @GetMapping("/fragments/fooHtml")
-    public String deleteTweet(@RequestParam("id") Optional<Long> id){
-        tweetService.deleteAllInfo(id.get());
+    public String deleteTweet(@RequestParam("id") Optional<Long> id, @RequestParam Optional<Boolean> changeSentiment,
+                              @RequestParam Optional<String> sentiment){
+        if(changeSentiment.get()){
+            tweetService.changeSentiment(sentiment.get(),id.get());
+        }else{
+            tweetService.deleteAllInfo(id.get());
+        }
         return "tweet/searchIndex";
     }
